+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @biconomy/sdk
 
+## 0.0.28
+
+### Patch Changes
+
+- Sudo fallback for empty rules
+
+## 0.0.27
+
+### Patch Changes
+
+- useRegistry false attestation fix
+
 ## 0.0.26
 
 ### Patch Changes